+++ component/src/main/java/io/siddhi/extension/io/file/FileSource.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.io.file;
+package io.siddhi.extension.io.file;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
@@ -33,16 +33,16 @@ import io.siddhi.core.util.config.ConfigReader;
 import io.siddhi.core.util.snapshot.state.State;
 import io.siddhi.core.util.snapshot.state.StateFactory;
 import io.siddhi.core.util.transport.OptionHolder;
+import io.siddhi.extension.io.file.listeners.FileSystemListener;
+import io.siddhi.extension.io.file.processors.FileProcessor;
+import io.siddhi.extension.io.file.util.Constants;
+import io.siddhi.extension.io.file.util.FileSourceConfiguration;
+import io.siddhi.extension.io.file.util.FileSourceServiceProvider;
+import io.siddhi.extension.io.file.util.VFSClientConnectorCallback;
 import org.apache.log4j.Logger;
 import org.wso2.carbon.messaging.ServerConnector;
 import org.wso2.carbon.messaging.exceptions.ClientConnectorException;
 import org.wso2.carbon.messaging.exceptions.ServerConnectorException;
-import org.wso2.extension.siddhi.io.file.listeners.FileSystemListener;
-import org.wso2.extension.siddhi.io.file.processors.FileProcessor;
-import org.wso2.extension.siddhi.io.file.util.Constants;
-import org.wso2.extension.siddhi.io.file.util.FileSourceConfiguration;
-import org.wso2.extension.siddhi.io.file.util.FileSourceServiceProvider;
-import org.wso2.extension.siddhi.io.file.util.VFSClientConnectorCallback;
 import org.wso2.transport.file.connector.sender.VFSClientConnector;
 import org.wso2.transport.file.connector.server.FileServerConnector;
 import org.wso2.transport.file.connector.server.FileServerConnectorProvider;
