+++ component/src/test/java/io/siddhi/extension/io/file/FileSourceLineModeTestCase.java
@@ -16,11 +16,12 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.io.file;
+package io.siddhi.extension.io.file;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
 import io.siddhi.core.event.Event;
+import io.siddhi.core.exception.CannotRestoreSiddhiAppStateException;
 import io.siddhi.core.exception.SiddhiAppCreationException;
 import io.siddhi.core.stream.output.StreamCallback;
 import io.siddhi.core.util.EventPrinter;
@@ -42,26 +43,25 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Test cases for siddhi-io-file source.
+ * mode = line.
  * */
-public class FileSourceRegexModeTestCase {
-    private static final Logger log = Logger.getLogger(FileSourceRegexModeTestCase.class);
+public class FileSourceLineModeTestCase {
+    private static final Logger log = Logger.getLogger(FileSourceLineModeTestCase.class);
     private AtomicInteger count = new AtomicInteger();
     private int waitTime = 2000;
     private int timeout = 30000;
-
     private String dirUri, moveAfterProcessDir;
     private File sourceRoot, newRoot, movedFiles;
-    
+
     @BeforeClass
     public void init() {
-        ClassLoader classLoader = FileSourceRegexModeTestCase.class.getClassLoader();
+        ClassLoader classLoader = FileSourceLineModeTestCase.class.getClassLoader();
         String rootPath = classLoader.getResource("files").getFile();
         sourceRoot = new File(rootPath + "/repo");
         dirUri = rootPath + "/new";
         newRoot = new File(dirUri);
         moveAfterProcessDir = rootPath + "/moved_files";
     }
-
     @BeforeMethod
     public void doBeforeMethod() {
         count.set(0);
@@ -76,7 +76,6 @@ public class FileSourceRegexModeTestCase {
                     " which are required for tests. Hence aborting tests.", e);
         }
     }
-
     @AfterMethod
     public void doAfterMethod() {
         try {
@@ -88,20 +87,18 @@ public class FileSourceRegexModeTestCase {
     }
 
     /**
-     * Test cases for 'mode = 'regex'.
-     * */
+     * Test cases for 'mode = line'.
+     */
 
     @Test
     public void siddhiIoFileTest1() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 1");
+        log.info("test SiddhiIoFile [mode=line] Test 1");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/json', " +
-                "begin.regex='(\\{)', " +
-                "end.regex='(}})', " +
-                "tailing='false', " +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/json', " +
                 "action.after.process='delete', " +
+                "tailing='false', " +
                 "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
@@ -113,14 +110,12 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
+                int n = count.getAndIncrement() % 5;
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -144,14 +139,10 @@ public class FileSourceRegexModeTestCase {
                 }
             }
         });
-
         siddhiAppRuntime.start();
-
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(dirUri + "/regex/json");
+        File file = new File(dirUri + "/line/json");
         AssertJUnit.assertEquals(0, file.list().length);
-
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
@@ -159,17 +150,17 @@ public class FileSourceRegexModeTestCase {
 
     @Test
     public void siddhiIoFileTest2() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 2");
+        log.info("test SiddhiIoFile [mode=line] Test 2");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/json', " +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/json', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
                 "tailing='false', " +
-                "action.after.process='delete', " +
                 "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
-
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -177,14 +168,12 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
+                int n = count.getAndIncrement() % 5;
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -208,14 +197,12 @@ public class FileSourceRegexModeTestCase {
                 }
             }
         });
-
         siddhiAppRuntime.start();
-
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(dirUri + "/regex/json");
+        File file = new File(dirUri + "/line/json");
         AssertJUnit.assertEquals(0, file.list().length);
-
+        File movedDir = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, movedDir.list().length);
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
@@ -223,339 +210,15 @@ public class FileSourceRegexModeTestCase {
 
     @Test
     public void siddhiIoFileTest3() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 3");
+        log.info("test SiddhiIoFile [mode=line] Test 3");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
-                "end.regex='(</events>)', " +
-                "tailing='false', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest4() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 4");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/json', " +
-                "begin.regex='(\\{)', " +
-                "end.regex='(}})', " +
-                "tailing='false', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
-                "@map(type='json'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest5() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 5");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
-                "tailing='false', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-        Thread.sleep(1000);
-
-        File file = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest6() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 6");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml_single/', " +
-                "begin.regex='(<events>)', " +
-                "tailing='false', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = events.length;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 1:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 8, count, timeout);
-
-        File file = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 8, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest7() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 7");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "end.regex='(</events>)', " +
-                "tailing='false', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 8, count, timeout);
-
-        File file = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest8() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 8");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
-                "end.regex='(</events>)', " +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/xml', " +
                 "tailing='true', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
-
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -563,7 +226,6 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
@@ -599,37 +261,33 @@ public class FileSourceRegexModeTestCase {
                 }
             }
         });
-
         Thread t1 = new Thread(new Runnable() {
             public void run() {
                 siddhiAppRuntime.start();
             }
         });
-
         t1.start();
-
         SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-
         Thread t2 = new Thread(new Runnable() {
             @Override
             public void run() {
-                File file = new File(dirUri + "/regex/xml/xml_logs (3rd copy).txt");
+                File file = new File(dirUri + "/line/xml/xml_line (3rd copy).txt");
                 try {
                     StringBuilder sb = new StringBuilder();
-                    sb.append("<events>\n")
-                            .append("<event>\n")
-                            .append("<symbol>").append("GOOGLE").append("</symbol>\n")
-                            .append("<price>").append("100").append("</price>\n")
-                            .append("<volume>").append("1000").append("</volume>\n")
-                            .append("</event>\n")
+                    sb.append("<events>")
+                            .append("<event>")
+                            .append("<symbol>").append("GOOGLE").append("</symbol>")
+                            .append("<price>").append("100").append("</price>")
+                            .append("<volume>").append("1000").append("</volume>")
+                            .append("</event>")
                             .append("</events>\n");
-                    sb.append("<events>\n")
-                            .append("<event>\n")
-                            .append("<symbol>").append("YAHOO").append("</symbol>\n")
-                            .append("<price>").append("200").append("</price>\n")
-                            .append("<volume>").append("2000").append("</volume>\n")
-                            .append("</event>\n")
-                            .append("</events>\n");
+                    sb.append("<events>")
+                            .append("<event>")
+                            .append("<symbol>").append("YAHOO").append("</symbol>")
+                            .append("<price>").append("200").append("</price>")
+                            .append("<volume>").append("2000").append("</volume>")
+                            .append("</event>")
+                            .append("</events>");
                     BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
                     bufferedWriter.write(sb.toString());
                     bufferedWriter.newLine();
@@ -642,26 +300,117 @@ public class FileSourceRegexModeTestCase {
         });
         t2.start();
         SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 7, count.get());
+        siddhiAppRuntime.shutdown();
+    }
+    @Test
+    public void siddhiIoFileTest4() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 4");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/xml', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
 
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.incrementAndGet();
+                for (Event event : events) {
+                    switch (n) {
+                        case 1:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 5:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        case 6:
+                            AssertJUnit.assertEquals(1000L, event.getData(2));
+                            break;
+                        case 7:
+                            AssertJUnit.assertEquals(2000L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+        Thread t1 = new Thread(new Runnable() {
+            public void run() {
+                siddhiAppRuntime.start();
+            }
+        });
+        t1.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
+        Thread t2 = new Thread(new Runnable() {
+            @Override
+            public void run() {
+                File file = new File(dirUri + "/line/xml/xml_line (3rd copy).txt");
+                try {
+                    StringBuilder sb = new StringBuilder();
+                    sb.append("<events>")
+                            .append("<event>")
+                            .append("<symbol>").append("GOOGLE").append("</symbol>")
+                            .append("<price>").append("100").append("</price>")
+                            .append("<volume>").append("1000").append("</volume>")
+                            .append("</event>")
+                            .append("</events>\n");
+                    sb.append("<events>")
+                            .append("<event>")
+                            .append("<symbol>").append("YAHOO").append("</symbol>")
+                            .append("<price>").append("200").append("</price>")
+                            .append("<volume>").append("2000").append("</volume>")
+                            .append("</event>")
+                            .append("</events>");
+                    BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
+                    bufferedWriter.write(sb.toString());
+                    bufferedWriter.newLine();
+                    bufferedWriter.flush();
+                    bufferedWriter.close();
+                } catch (IOException e) {
+                    log.error(e.getMessage());
+                }
+            }
+        });
+        t2.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
         //assert event count
         AssertJUnit.assertEquals("Number of events", 7, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
-    public void siddhiIoFileTest9() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 9");
+    public void siddhiIoFileTest5() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 5");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/json', " +
                 "tailing='false', " +
-                "action.after.process='delete', " +
-                "@map(type='xml'))" +
+                "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
-
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -669,14 +418,12 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
+                int n = count.getAndIncrement() % 5;
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -700,135 +447,27 @@ public class FileSourceRegexModeTestCase {
                 }
             }
         });
-
         siddhiAppRuntime.start();
-
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(dirUri + "/regex/xml");
+        File file = new File(dirUri + "/line/json");
         AssertJUnit.assertEquals(0, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test (expectedExceptions = SiddhiAppCreationException.class)
-    public void siddhiIoFileTest10() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 10");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "begin.regex='(<events>)', " +
-                "tailing='false', " +
-                "action.after.process='delete', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        Thread.sleep(2000);
-
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest11() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 11");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
-                "end.regex='(</events>)', " +
-                "tailing='false', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(dirUri + "/regex/xml");
-        AssertJUnit.assertEquals(0, file.list().length);
-
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void siddhiIoFileTest12() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 12");
+    public void siddhiIoFileTest6() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 6");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
-                "end.regex='(</events>)', " +
-                "action.after.process='move' ," +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/json', " +
+                "action.after.process='move', " +
                 "tailing='false', " +
-                "@map(type='xml'))" +
+                "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
-
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -836,14 +475,156 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
+            }
+        });
+        siddhiAppRuntime.start();
+        Thread.sleep(1000);
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void siddhiIoFileTest7() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 7");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/json_invalid_path', " +
+                "action.after.process='move', " +
+                "tailing='false', " +
+                "@map(type='json'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+            }
+        });
+        siddhiAppRuntime.start();
+        Thread.sleep(1000);
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test
+    public void siddhiIoFileTest8() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 8");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/invalid', " +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.incrementAndGet();
+                for (Event event : events) {
+                    switch (n) {
+                        case 1:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+        siddhiAppRuntime.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 3, count, timeout);
+        File file = new File(dirUri + "/line/invalid");
+        AssertJUnit.assertEquals(0, file.list().length);
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 3, count.get());
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void siddhiIoFileTest9() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 9");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/invalid', " +
+                "tailing='false'," +
+                "action.after.process='delete', " +
+                "move.after.process='file://abc/def/', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+            }
+        });
+        siddhiAppRuntime.start();
+        Thread.sleep(1000);
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test
+    public void siddhiIoFileTest10() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 10");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
+                "action.after.process='delete', " +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.getAndIncrement();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -867,14 +648,132 @@ public class FileSourceRegexModeTestCase {
                 }
             }
         });
-
         siddhiAppRuntime.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
+        File file = new File(dirUri + "/line/xml/");
+        AssertJUnit.assertEquals(7, file.list().length);
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 5, count.get());
+        siddhiAppRuntime.shutdown();
+    }
 
+    @Test
+    public void siddhiIoFileTest11() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 11");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/xml', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.getAndIncrement() % 5;
+                for (Event event : events) {
+                    switch (n) {
+                        case 0:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 1:
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+        siddhiAppRuntime.start();
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(dirUri + "/regex/xml");
+        File file = new File(dirUri + "/line/xml");
         AssertJUnit.assertEquals(0, file.list().length);
+        File movedDir = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, movedDir.list().length);
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
+        siddhiAppRuntime.shutdown();
+    }
 
+    @Test
+    public void siddhiIoFileTest12() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 12");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/text', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "tailing='false', " +
+                "@map(type='text',fail.on.missing.attribute='false', " +
+                "regex.A='(\\w+),([-.0-9]+),([-.0-9]+)', event.grouping.enabled='false', " +
+                "@attributes(symbol = 'A[1]', price = 'A[2]', volume = 'A[3]')))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.getAndIncrement() % 5;
+                for (Event event : events) {
+                    switch (n) {
+                        case 0:
+                            AssertJUnit.assertEquals(100L, event.getData(2));
+                            break;
+                        case 1:
+                            AssertJUnit.assertEquals(200L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(300L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(400L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(500L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+        siddhiAppRuntime.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
+        File file = new File(dirUri + "/line/text");
+        AssertJUnit.assertEquals(0, file.list().length);
+        File movedDir = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, movedDir.list().length);
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
@@ -882,18 +781,15 @@ public class FileSourceRegexModeTestCase {
 
     @Test
     public void siddhiIoFileTest13() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 13");
+        log.info("test SiddhiIoFile [mode=line] Test 13");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/invalid', " +
-                "begin.regex='(<events>)', " +
-                "end.regex='(</events>)', " +
-                "tailing='false', " +
+                "@source(type='file', mode='line'," +
+                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
+                "tailing='true', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
-
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -901,7 +797,6 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
@@ -914,10 +809,22 @@ public class FileSourceRegexModeTestCase {
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
                         case 2:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
                         case 3:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 5:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        case 6:
+                            AssertJUnit.assertEquals(1000L, event.getData(2));
+                            break;
+                        case 7:
+                            AssertJUnit.assertEquals(2000L, event.getData(2));
                             break;
                         default:
                             AssertJUnit.fail("More events received than expected.");
@@ -925,32 +832,65 @@ public class FileSourceRegexModeTestCase {
                 }
             }
         });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 3, count, timeout);
-
-        File file = new File(dirUri + "/regex/invalid");
-        AssertJUnit.assertEquals(0, file.list().length);
-
+        Thread t1 = new Thread(new Runnable() {
+            public void run() {
+                siddhiAppRuntime.start();
+            }
+        });
+        t1.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
+        Thread t2 = new Thread(new Runnable() {
+            @Override
+            public void run() {
+                File file = new File(dirUri + "/line/xml/xml_line.txt");
+                try {
+                    StringBuilder sb = new StringBuilder();
+                    sb.append("<events>")
+                            .append("<event>")
+                            .append("<symbol>").append("GOOGLE").append("</symbol>")
+                            .append("<price>").append("100").append("</price>")
+                            .append("<volume>").append("1000").append("</volume>")
+                            .append("</event>")
+                            .append("</events>\n");
+                    sb.append("<events>")
+                            .append("<event>")
+                            .append("<symbol>").append("YAHOO").append("</symbol>")
+                            .append("<price>").append("200").append("</price>")
+                            .append("<volume>").append("2000").append("</volume>")
+                            .append("</event>")
+                            .append("</events>");
+                    BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
+                    bufferedWriter.write(sb.toString());
+                    bufferedWriter.newLine();
+                    bufferedWriter.flush();
+                    bufferedWriter.close();
+                } catch (IOException e) {
+                    log.error(e.getMessage());
+                }
+            }
+        });
+        t2.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 3, count.get());
+        AssertJUnit.assertEquals("Number of events", 7, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
     public void siddhiIoFileTest14() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 14");
+        log.info("test SiddhiIoFile [mode=line] Test 14");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/invalid', " +
-                "begin.regex='(<events>)', " +
+                "@source(type='file', mode='line'," +
+                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "/xml_line.txt', " +
+                "action.after.failure='move', " +
+                "move.after.failure='file:/" + moveAfterProcessDir + "/xml_line.txt', " +
                 "tailing='false', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
-
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -958,6 +898,91 @@ public class FileSourceRegexModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.getAndIncrement();
+                for (Event event : events) {
+                    switch (n) {
+                        case 0:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 1:
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+        siddhiAppRuntime.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
+        File file = new File(dirUri + "/line/xml/");
+        AssertJUnit.assertEquals(7, file.list().length);
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 5, count.get());
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void siddhiIoFileTest15() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode=line] Test 15");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "dir.uri='file:/" + dirUri + "/line/text', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "tailing='true', " +
+                "@map(type='text',fail.on.missing.attribute='false', " +
+                "regex.A='(\\w+),([-.0-9]+),([-.0-9]+)', event.grouping.enabled='false', " +
+                "@attributes(symbol = 'A[1]', price = 'A[2]', volume = 'A[3]')))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
+        siddhiAppRuntime.start();
+        Thread.sleep(1000);
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test
+    public void siddhiIoFileTest16() throws InterruptedException, CannotRestoreSiddhiAppStateException {
+        log.info("test SiddhiIoFile [mode=line] Test 16");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='line'," +
+                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
+                "tailing='true', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        SiddhiAppRuntime siddhiAppRuntime2 = siddhiManager.createSiddhiAppRuntime(streams + query);
 
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
@@ -971,181 +996,83 @@ public class FileSourceRegexModeTestCase {
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
                         case 2:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
                         case 3:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 5:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
                             break;
+
                         default:
-                            AssertJUnit.fail("More events received than expected.");
+                            AssertJUnit.fail();
+
                     }
                 }
             }
         });
 
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 3, count, timeout);
-
-        File file = new File(dirUri + "/regex/invalid");
-        AssertJUnit.assertEquals(0, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 3, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest15() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 15");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "end.regex='(</events>)', " +
-                "tailing='false', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+        siddhiAppRuntime2.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
+                int n = count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                        case 6:
+                            AssertJUnit.assertEquals(1000L, event.getData(2));
                             break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                        case 7:
+                            AssertJUnit.assertEquals(2000L, event.getData(2));
                             break;
                         default:
-                            AssertJUnit.fail("More events received than expected.");
+                            AssertJUnit.fail();
+
                     }
                 }
             }
         });
-
         siddhiAppRuntime.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
+        byte[] snapshot = siddhiAppRuntime.snapshot();
+        siddhiAppRuntime.shutdown();
 
-        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
+        File file = new File(dirUri + "/line/xml/xml_line.txt");
+        try {
+            StringBuilder sb = new StringBuilder();
+            sb.append("<events>")
+                    .append("<event>")
+                    .append("<symbol>").append("GOOGLE").append("</symbol>")
+                    .append("<price>").append("100").append("</price>")
+                    .append("<volume>").append("1000").append("</volume>")
+                    .append("</event>")
+                    .append("</events>\n");
+            sb.append("<events>")
+                    .append("<event>")
+                    .append("<symbol>").append("YAHOO").append("</symbol>")
+                    .append("<price>").append("200").append("</price>")
+                    .append("<volume>").append("2000").append("</volume>")
+                    .append("</event>")
+                    .append("</events>");
+            BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
+            bufferedWriter.write(sb.toString());
+            bufferedWriter.newLine();
+            bufferedWriter.flush();
+            bufferedWriter.close();
+        } catch (IOException e) {
+            log.error(e.getMessage());
+        }
 
-        File file = new File(dirUri + "/regex/xml");
-        AssertJUnit.assertEquals(0, file.list().length);
-
+        siddhiAppRuntime2.restore(snapshot);
+        siddhiAppRuntime2.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
+        AssertJUnit.assertEquals("Number of events", 7, count.get());
         siddhiAppRuntime.shutdown();
     }
-
-    @Test
-    public void siddhiIoFileTest16() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 16");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(<events>)', " +
-                "tailing='false', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.intValue() % 5;
-                count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 2:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 4:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-
-        siddhiAppRuntime.start();
-
-        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-
-        File file = new File(dirUri + "/regex/xml");
-        AssertJUnit.assertEquals(0, file.list().length);
-
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 40, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void siddhiIoFileTest17() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode = regex] 17");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='regex'," +
-                "dir.uri='file:/" + dirUri + "/regex/xml', " +
-                "begin.regex='(*<events>)', " +
-                "tailing='false', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.start();
-        Thread.sleep(100);
-        siddhiAppRuntime.shutdown();
-    }
-
 }
