+++ component/src/test/java/io/siddhi/extension/io/file/FileSourceRegexModeTestCase.java
@@ -16,12 +16,11 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.io.file;
+package io.siddhi.extension.io.file;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
 import io.siddhi.core.event.Event;
-import io.siddhi.core.exception.CannotRestoreSiddhiAppStateException;
 import io.siddhi.core.exception.SiddhiAppCreationException;
 import io.siddhi.core.stream.output.StreamCallback;
 import io.siddhi.core.util.EventPrinter;
@@ -43,25 +42,26 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Test cases for siddhi-io-file source.
- * mode = line.
  * */
-public class FileSourceLineModeTestCase {
-    private static final Logger log = Logger.getLogger(FileSourceLineModeTestCase.class);
+public class FileSourceRegexModeTestCase {
+    private static final Logger log = Logger.getLogger(FileSourceRegexModeTestCase.class);
     private AtomicInteger count = new AtomicInteger();
     private int waitTime = 2000;
     private int timeout = 30000;
+
     private String dirUri, moveAfterProcessDir;
     private File sourceRoot, newRoot, movedFiles;
-
+    
     @BeforeClass
     public void init() {
-        ClassLoader classLoader = FileSourceLineModeTestCase.class.getClassLoader();
+        ClassLoader classLoader = FileSourceRegexModeTestCase.class.getClassLoader();
         String rootPath = classLoader.getResource("files").getFile();
         sourceRoot = new File(rootPath + "/repo");
         dirUri = rootPath + "/new";
         newRoot = new File(dirUri);
         moveAfterProcessDir = rootPath + "/moved_files";
     }
+
     @BeforeMethod
     public void doBeforeMethod() {
         count.set(0);
@@ -76,6 +76,7 @@ public class FileSourceLineModeTestCase {
                     " which are required for tests. Hence aborting tests.", e);
         }
     }
+
     @AfterMethod
     public void doAfterMethod() {
         try {
@@ -87,18 +88,20 @@ public class FileSourceLineModeTestCase {
     }
 
     /**
-     * Test cases for 'mode = line'.
-     */
+     * Test cases for 'mode = 'regex'.
+     * */
 
     @Test
     public void siddhiIoFileTest1() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 1");
+        log.info("test SiddhiIoFile [mode = regex] 1");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/json', " +
-                "action.after.process='delete', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/json', " +
+                "begin.regex='(\\{)', " +
+                "end.regex='(}})', " +
                 "tailing='false', " +
+                "action.after.process='delete', " +
                 "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
@@ -110,12 +113,14 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.getAndIncrement() % 5;
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -139,10 +144,14 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
+
         siddhiAppRuntime.start();
+
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-        File file = new File(dirUri + "/line/json");
+
+        File file = new File(dirUri + "/regex/json");
         AssertJUnit.assertEquals(0, file.list().length);
+
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
@@ -150,17 +159,17 @@ public class FileSourceLineModeTestCase {
 
     @Test
     public void siddhiIoFileTest2() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 2");
+        log.info("test SiddhiIoFile [mode = regex] 2");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/json', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/json', " +
                 "tailing='false', " +
+                "action.after.process='delete', " +
                 "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -168,12 +177,14 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.getAndIncrement() % 5;
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -197,12 +208,14 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
+
         siddhiAppRuntime.start();
+
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-        File file = new File(dirUri + "/line/json");
+
+        File file = new File(dirUri + "/regex/json");
         AssertJUnit.assertEquals(0, file.list().length);
-        File movedDir = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, movedDir.list().length);
+
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
@@ -210,15 +223,20 @@ public class FileSourceLineModeTestCase {
 
     @Test
     public void siddhiIoFileTest3() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 3");
+        log.info("test SiddhiIoFile [mode = regex] 3");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/xml', " +
-                "tailing='true', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
+                "end.regex='(</events>)', " +
+                "tailing='false', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -226,34 +244,30 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.incrementAndGet();
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
-                        case 1:
+                        case 0:
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
-                        case 2:
+                        case 1:
                             AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
-                        case 3:
+                        case 2:
                             AssertJUnit.assertEquals(10002L, event.getData(2));
                             break;
-                        case 4:
+                        case 3:
                             AssertJUnit.assertEquals(10003L, event.getData(2));
                             break;
-                        case 5:
+                        case 4:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        case 6:
-                            AssertJUnit.assertEquals(1000L, event.getData(2));
-                            break;
-                        case 7:
-                            AssertJUnit.assertEquals(2000L, event.getData(2));
                             break;
                         default:
                             AssertJUnit.fail("More events received than expected.");
@@ -261,59 +275,35 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
-        Thread t1 = new Thread(new Runnable() {
-            public void run() {
-                siddhiAppRuntime.start();
-            }
-        });
-        t1.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-        Thread t2 = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                File file = new File(dirUri + "/line/xml/xml_line (3rd copy).txt");
-                try {
-                    StringBuilder sb = new StringBuilder();
-                    sb.append("<events>")
-                            .append("<event>")
-                            .append("<symbol>").append("GOOGLE").append("</symbol>")
-                            .append("<price>").append("100").append("</price>")
-                            .append("<volume>").append("1000").append("</volume>")
-                            .append("</event>")
-                            .append("</events>\n");
-                    sb.append("<events>")
-                            .append("<event>")
-                            .append("<symbol>").append("YAHOO").append("</symbol>")
-                            .append("<price>").append("200").append("</price>")
-                            .append("<volume>").append("2000").append("</volume>")
-                            .append("</event>")
-                            .append("</events>");
-                    BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
-                    bufferedWriter.write(sb.toString());
-                    bufferedWriter.newLine();
-                    bufferedWriter.flush();
-                    bufferedWriter.close();
-                } catch (IOException e) {
-                    log.error(e.getMessage());
-                }
-            }
-        });
-        t2.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
+
+        siddhiAppRuntime.start();
+
+        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
+
+        File file = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, file.list().length);
+
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 7, count.get());
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
+
     @Test
     public void siddhiIoFileTest4() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 4");
+        log.info("test SiddhiIoFile [mode = regex] 4");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/xml', " +
-                "@map(type='xml'))" +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/json', " +
+                "begin.regex='(\\{)', " +
+                "end.regex='(}})', " +
+                "tailing='false', " +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@map(type='json'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -321,34 +311,30 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.incrementAndGet();
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
-                        case 1:
+                        case 0:
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
-                        case 2:
+                        case 1:
                             AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
-                        case 3:
+                        case 2:
                             AssertJUnit.assertEquals(10002L, event.getData(2));
                             break;
-                        case 4:
+                        case 3:
                             AssertJUnit.assertEquals(10003L, event.getData(2));
                             break;
-                        case 5:
+                        case 4:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        case 6:
-                            AssertJUnit.assertEquals(1000L, event.getData(2));
-                            break;
-                        case 7:
-                            AssertJUnit.assertEquals(2000L, event.getData(2));
                             break;
                         default:
                             AssertJUnit.fail("More events received than expected.");
@@ -356,61 +342,34 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
-        Thread t1 = new Thread(new Runnable() {
-            public void run() {
-                siddhiAppRuntime.start();
-            }
-        });
-        t1.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-        Thread t2 = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                File file = new File(dirUri + "/line/xml/xml_line (3rd copy).txt");
-                try {
-                    StringBuilder sb = new StringBuilder();
-                    sb.append("<events>")
-                            .append("<event>")
-                            .append("<symbol>").append("GOOGLE").append("</symbol>")
-                            .append("<price>").append("100").append("</price>")
-                            .append("<volume>").append("1000").append("</volume>")
-                            .append("</event>")
-                            .append("</events>\n");
-                    sb.append("<events>")
-                            .append("<event>")
-                            .append("<symbol>").append("YAHOO").append("</symbol>")
-                            .append("<price>").append("200").append("</price>")
-                            .append("<volume>").append("2000").append("</volume>")
-                            .append("</event>")
-                            .append("</events>");
-                    BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
-                    bufferedWriter.write(sb.toString());
-                    bufferedWriter.newLine();
-                    bufferedWriter.flush();
-                    bufferedWriter.close();
-                } catch (IOException e) {
-                    log.error(e.getMessage());
-                }
-            }
-        });
-        t2.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
+
+        siddhiAppRuntime.start();
+
+        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
+
+        File file = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, file.list().length);
+
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 7, count.get());
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
     public void siddhiIoFileTest5() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 5");
+        log.info("test SiddhiIoFile [mode = regex] 5");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/json', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
                 "tailing='false', " +
-                "@map(type='json'))" +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -418,12 +377,14 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.getAndIncrement() % 5;
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -447,27 +408,35 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
+
         siddhiAppRuntime.start();
+
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-        File file = new File(dirUri + "/line/json");
-        AssertJUnit.assertEquals(0, file.list().length);
+        Thread.sleep(1000);
+
+        File file = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, file.list().length);
+
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
 
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    @Test
     public void siddhiIoFileTest6() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 6");
+        log.info("test SiddhiIoFile [mode = regex] 6");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/json', " +
-                "action.after.process='move', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml_single/', " +
+                "begin.regex='(<events>)', " +
                 "tailing='false', " +
-                "@map(type='json'))" +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -475,30 +444,53 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
+                int n = events.length;
+                count.incrementAndGet();
+                for (Event event : events) {
+                    switch (n) {
+                        case 1:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
             }
         });
+
         siddhiAppRuntime.start();
-        Thread.sleep(1000);
+
+        SiddhiTestHelper.waitForEvents(waitTime, 8, count, timeout);
+
+        File file = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, file.list().length);
+
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 8, count.get());
         siddhiAppRuntime.shutdown();
     }
 
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    @Test
     public void siddhiIoFileTest7() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 7");
+        log.info("test SiddhiIoFile [mode = regex] 7");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/json_invalid_path', " +
-                "action.after.process='move', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "end.regex='(</events>)', " +
                 "tailing='false', " +
-                "@map(type='json'))" +
+                "action.after.process='move', " +
+                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -506,29 +498,64 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
+                for (Event event : events) {
+                    switch (n) {
+                        case 0:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 1:
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
             }
         });
+
         siddhiAppRuntime.start();
-        Thread.sleep(1000);
+
+        SiddhiTestHelper.waitForEvents(waitTime, 8, count, timeout);
+
+        File file = new File(moveAfterProcessDir);
+        AssertJUnit.assertEquals(8, file.list().length);
+
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
     public void siddhiIoFileTest8() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 8");
+        log.info("test SiddhiIoFile [mode = regex] 8");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/invalid', " +
-                "tailing='false', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
+                "end.regex='(</events>)', " +
+                "tailing='true', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -536,6 +563,7 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
@@ -548,99 +576,22 @@ public class FileSourceLineModeTestCase {
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
                         case 2:
-                            AssertJUnit.assertEquals(10003L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail("More events received than expected.");
-                    }
-                }
-            }
-        });
-        siddhiAppRuntime.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 3, count, timeout);
-        File file = new File(dirUri + "/line/invalid");
-        AssertJUnit.assertEquals(0, file.list().length);
-        //assert event count
-        AssertJUnit.assertEquals("Number of events", 3, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void siddhiIoFileTest9() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 9");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/invalid', " +
-                "tailing='false'," +
-                "action.after.process='delete', " +
-                "move.after.process='file://abc/def/', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-            }
-        });
-        siddhiAppRuntime.start();
-        Thread.sleep(1000);
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test
-    public void siddhiIoFileTest10() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 10");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
-                "action.after.process='delete', " +
-                "tailing='false', " +
-                "@map(type='xml'))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.getAndIncrement();
-                for (Event event : events) {
-                    switch (n) {
-                        case 0:
-                            AssertJUnit.assertEquals(10000L, event.getData(2));
-                            break;
-                        case 1:
                             AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
-                        case 2:
+                        case 3:
                             AssertJUnit.assertEquals(10002L, event.getData(2));
                             break;
-                        case 3:
+                        case 4:
                             AssertJUnit.assertEquals(10003L, event.getData(2));
                             break;
-                        case 4:
+                        case 5:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        case 6:
+                            AssertJUnit.assertEquals(1000L, event.getData(2));
+                            break;
+                        case 7:
+                            AssertJUnit.assertEquals(2000L, event.getData(2));
                             break;
                         default:
                             AssertJUnit.fail("More events received than expected.");
@@ -648,28 +599,69 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
-        siddhiAppRuntime.start();
+
+        Thread t1 = new Thread(new Runnable() {
+            public void run() {
+                siddhiAppRuntime.start();
+            }
+        });
+
+        t1.start();
+
         SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-        File file = new File(dirUri + "/line/xml/");
-        AssertJUnit.assertEquals(7, file.list().length);
+
+        Thread t2 = new Thread(new Runnable() {
+            @Override
+            public void run() {
+                File file = new File(dirUri + "/regex/xml/xml_logs (3rd copy).txt");
+                try {
+                    StringBuilder sb = new StringBuilder();
+                    sb.append("<events>\n")
+                            .append("<event>\n")
+                            .append("<symbol>").append("GOOGLE").append("</symbol>\n")
+                            .append("<price>").append("100").append("</price>\n")
+                            .append("<volume>").append("1000").append("</volume>\n")
+                            .append("</event>\n")
+                            .append("</events>\n");
+                    sb.append("<events>\n")
+                            .append("<event>\n")
+                            .append("<symbol>").append("YAHOO").append("</symbol>\n")
+                            .append("<price>").append("200").append("</price>\n")
+                            .append("<volume>").append("2000").append("</volume>\n")
+                            .append("</event>\n")
+                            .append("</events>\n");
+                    BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
+                    bufferedWriter.write(sb.toString());
+                    bufferedWriter.newLine();
+                    bufferedWriter.flush();
+                    bufferedWriter.close();
+                } catch (IOException e) {
+                    log.error(e.getMessage());
+                }
+            }
+        });
+        t2.start();
+        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
+
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 5, count.get());
+        AssertJUnit.assertEquals("Number of events", 7, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
-    public void siddhiIoFileTest11() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 11");
+    public void siddhiIoFileTest9() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 9");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/xml', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
                 "tailing='false', " +
+                "action.after.process='delete', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -677,12 +669,14 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.getAndIncrement() % 5;
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -706,32 +700,32 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
+
         siddhiAppRuntime.start();
+
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-        File file = new File(dirUri + "/line/xml");
+
+        File file = new File(dirUri + "/regex/xml");
         AssertJUnit.assertEquals(0, file.list().length);
-        File movedDir = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, movedDir.list().length);
+
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
 
-    @Test
-    public void siddhiIoFileTest12() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 12");
+    @Test (expectedExceptions = SiddhiAppCreationException.class)
+    public void siddhiIoFileTest10() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 10");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/text', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
+                "@source(type='file', mode='regex'," +
+                "begin.regex='(<events>)', " +
                 "tailing='false', " +
-                "@map(type='text',fail.on.missing.attribute='false', " +
-                "regex.A='(\\w+),([-.0-9]+),([-.0-9]+)', event.grouping.enabled='false', " +
-                "@attributes(symbol = 'A[1]', price = 'A[2]', volume = 'A[3]')))" +
+                "action.after.process='delete', " +
+                "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -739,28 +733,67 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.getAndIncrement() % 5;
+            }
+        });
+
+        siddhiAppRuntime.start();
+
+        Thread.sleep(2000);
+
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test
+    public void siddhiIoFileTest11() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 11");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
+                "end.regex='(</events>)', " +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
-                            AssertJUnit.assertEquals(100L, event.getData(2));
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
                         case 1:
-                            AssertJUnit.assertEquals(200L, event.getData(2));
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
                         case 2:
-                            AssertJUnit.assertEquals(300L, event.getData(2));
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
                             break;
                         case 3:
-                            AssertJUnit.assertEquals(400L, event.getData(2));
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
                             break;
                         case 4:
-                            AssertJUnit.assertEquals(500L, event.getData(2));
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
                             break;
                         default:
                             AssertJUnit.fail("More events received than expected.");
@@ -768,12 +801,80 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
+
         siddhiAppRuntime.start();
+
         SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
-        File file = new File(dirUri + "/line/text");
+
+        File file = new File(dirUri + "/regex/xml");
         AssertJUnit.assertEquals(0, file.list().length);
-        File movedDir = new File(moveAfterProcessDir);
-        AssertJUnit.assertEquals(8, movedDir.list().length);
+
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void siddhiIoFileTest12() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 12");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
+                "end.regex='(</events>)', " +
+                "action.after.process='move' ," +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
+                for (Event event : events) {
+                    switch (n) {
+                        case 0:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 1:
+                            AssertJUnit.assertEquals(10001L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10002L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 4:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+
+        siddhiAppRuntime.start();
+
+        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
+
+        File file = new File(dirUri + "/regex/xml");
+        AssertJUnit.assertEquals(0, file.list().length);
+
         //assert event count
         AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
@@ -781,15 +882,18 @@ public class FileSourceLineModeTestCase {
 
     @Test
     public void siddhiIoFileTest13() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 13");
+        log.info("test SiddhiIoFile [mode = regex] 13");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
-                "tailing='true', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/invalid', " +
+                "begin.regex='(<events>)', " +
+                "end.regex='(</events>)', " +
+                "tailing='false', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -797,6 +901,7 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
@@ -809,22 +914,10 @@ public class FileSourceLineModeTestCase {
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
                         case 2:
-                            AssertJUnit.assertEquals(10001L, event.getData(2));
-                            break;
-                        case 3:
-                            AssertJUnit.assertEquals(10002L, event.getData(2));
-                            break;
-                        case 4:
                             AssertJUnit.assertEquals(10003L, event.getData(2));
                             break;
-                        case 5:
+                        case 3:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
-                            break;
-                        case 6:
-                            AssertJUnit.assertEquals(1000L, event.getData(2));
-                            break;
-                        case 7:
-                            AssertJUnit.assertEquals(2000L, event.getData(2));
                             break;
                         default:
                             AssertJUnit.fail("More events received than expected.");
@@ -832,65 +925,32 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
-        Thread t1 = new Thread(new Runnable() {
-            public void run() {
-                siddhiAppRuntime.start();
-            }
-        });
-        t1.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-        Thread t2 = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                File file = new File(dirUri + "/line/xml/xml_line.txt");
-                try {
-                    StringBuilder sb = new StringBuilder();
-                    sb.append("<events>")
-                            .append("<event>")
-                            .append("<symbol>").append("GOOGLE").append("</symbol>")
-                            .append("<price>").append("100").append("</price>")
-                            .append("<volume>").append("1000").append("</volume>")
-                            .append("</event>")
-                            .append("</events>\n");
-                    sb.append("<events>")
-                            .append("<event>")
-                            .append("<symbol>").append("YAHOO").append("</symbol>")
-                            .append("<price>").append("200").append("</price>")
-                            .append("<volume>").append("2000").append("</volume>")
-                            .append("</event>")
-                            .append("</events>");
-                    BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
-                    bufferedWriter.write(sb.toString());
-                    bufferedWriter.newLine();
-                    bufferedWriter.flush();
-                    bufferedWriter.close();
-                } catch (IOException e) {
-                    log.error(e.getMessage());
-                }
-            }
-        });
-        t2.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
+
+        siddhiAppRuntime.start();
+
+        SiddhiTestHelper.waitForEvents(waitTime, 3, count, timeout);
+
+        File file = new File(dirUri + "/regex/invalid");
+        AssertJUnit.assertEquals(0, file.list().length);
+
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 7, count.get());
+        AssertJUnit.assertEquals("Number of events", 3, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
     public void siddhiIoFileTest14() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 14");
+        log.info("test SiddhiIoFile [mode = regex] 14");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "/xml_line.txt', " +
-                "action.after.failure='move', " +
-                "move.after.failure='file:/" + moveAfterProcessDir + "/xml_line.txt', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/invalid', " +
+                "begin.regex='(<events>)', " +
                 "tailing='false', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -898,12 +958,71 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.getAndIncrement();
+                int n = count.incrementAndGet();
+                for (Event event : events) {
+                    switch (n) {
+                        case 1:
+                            AssertJUnit.assertEquals(10000L, event.getData(2));
+                            break;
+                        case 2:
+                            AssertJUnit.assertEquals(10003L, event.getData(2));
+                            break;
+                        case 3:
+                            AssertJUnit.assertEquals(10004L, event.getData(2));
+                            break;
+                        default:
+                            AssertJUnit.fail("More events received than expected.");
+                    }
+                }
+            }
+        });
+
+        siddhiAppRuntime.start();
+
+        SiddhiTestHelper.waitForEvents(waitTime, 3, count, timeout);
+
+        File file = new File(dirUri + "/regex/invalid");
+        AssertJUnit.assertEquals(0, file.list().length);
+
+        //assert event count
+        AssertJUnit.assertEquals("Number of events", 3, count.get());
+        siddhiAppRuntime.shutdown();
+    }
+
+    @Test
+    public void siddhiIoFileTest15() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 15");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "end.regex='(</events>)', " +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
+        siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
+
+            @Override
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
                         case 0:
@@ -927,54 +1046,32 @@ public class FileSourceLineModeTestCase {
                 }
             }
         });
+
         siddhiAppRuntime.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-        File file = new File(dirUri + "/line/xml/");
-        AssertJUnit.assertEquals(7, file.list().length);
+
+        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
+
+        File file = new File(dirUri + "/regex/xml");
+        AssertJUnit.assertEquals(0, file.list().length);
+
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 5, count.get());
-        siddhiAppRuntime.shutdown();
-    }
-
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void siddhiIoFileTest15() throws InterruptedException {
-        log.info("test SiddhiIoFile [mode=line] Test 15");
-        String streams = "" +
-                "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "dir.uri='file:/" + dirUri + "/line/text', " +
-                "action.after.process='move', " +
-                "move.after.process='file:/" + moveAfterProcessDir + "', " +
-                "tailing='true', " +
-                "@map(type='text',fail.on.missing.attribute='false', " +
-                "regex.A='(\\w+),([-.0-9]+),([-.0-9]+)', event.grouping.enabled='false', " +
-                "@attributes(symbol = 'A[1]', price = 'A[2]', volume = 'A[3]')))" +
-                "define stream FooStream (symbol string, price float, volume long); " +
-                "define stream BarStream (symbol string, price float, volume long); ";
-        String query = "" +
-                "from FooStream " +
-                "select * " +
-                "insert into BarStream; ";
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-
-        siddhiAppRuntime.start();
-        Thread.sleep(1000);
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
 
     @Test
-    public void siddhiIoFileTest16() throws InterruptedException, CannotRestoreSiddhiAppStateException {
-        log.info("test SiddhiIoFile [mode=line] Test 16");
+    public void siddhiIoFileTest16() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 16");
         String streams = "" +
                 "@App:name('TestSiddhiApp')" +
-                "@source(type='file', mode='line'," +
-                "file.uri='file:/" + dirUri + "/line/xml/xml_line.txt', " +
-                "tailing='true', " +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(<events>)', " +
+                "tailing='false', " +
                 "@map(type='xml'))" +
                 "define stream FooStream (symbol string, price float, volume long); " +
                 "define stream BarStream (symbol string, price float, volume long); ";
+
         String query = "" +
                 "from FooStream " +
                 "select * " +
@@ -982,97 +1079,73 @@ public class FileSourceLineModeTestCase {
 
         SiddhiManager siddhiManager = new SiddhiManager();
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        SiddhiAppRuntime siddhiAppRuntime2 = siddhiManager.createSiddhiAppRuntime(streams + query);
 
         siddhiAppRuntime.addCallback("BarStream", new StreamCallback() {
 
             @Override
             public void receive(Event[] events) {
                 EventPrinter.print(events);
-                int n = count.incrementAndGet();
+                int n = count.intValue() % 5;
+                count.incrementAndGet();
                 for (Event event : events) {
                     switch (n) {
-                        case 1:
+                        case 0:
                             AssertJUnit.assertEquals(10000L, event.getData(2));
                             break;
-                        case 2:
+                        case 1:
                             AssertJUnit.assertEquals(10001L, event.getData(2));
                             break;
-                        case 3:
+                        case 2:
                             AssertJUnit.assertEquals(10002L, event.getData(2));
                             break;
-                        case 4:
+                        case 3:
                             AssertJUnit.assertEquals(10003L, event.getData(2));
                             break;
-                        case 5:
+                        case 4:
                             AssertJUnit.assertEquals(10004L, event.getData(2));
                             break;
-
                         default:
-                            AssertJUnit.fail();
-
+                            AssertJUnit.fail("More events received than expected.");
                     }
                 }
             }
         });
 
-        siddhiAppRuntime2.addCallback("BarStream", new StreamCallback() {
-
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                int n = count.incrementAndGet();
-                for (Event event : events) {
-                    switch (n) {
-                        case 6:
-                            AssertJUnit.assertEquals(1000L, event.getData(2));
-                            break;
-                        case 7:
-                            AssertJUnit.assertEquals(2000L, event.getData(2));
-                            break;
-                        default:
-                            AssertJUnit.fail();
-
-                    }
-                }
-            }
-        });
         siddhiAppRuntime.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 5, count, timeout);
-        byte[] snapshot = siddhiAppRuntime.snapshot();
-        siddhiAppRuntime.shutdown();
 
-        File file = new File(dirUri + "/line/xml/xml_line.txt");
-        try {
-            StringBuilder sb = new StringBuilder();
-            sb.append("<events>")
-                    .append("<event>")
-                    .append("<symbol>").append("GOOGLE").append("</symbol>")
-                    .append("<price>").append("100").append("</price>")
-                    .append("<volume>").append("1000").append("</volume>")
-                    .append("</event>")
-                    .append("</events>\n");
-            sb.append("<events>")
-                    .append("<event>")
-                    .append("<symbol>").append("YAHOO").append("</symbol>")
-                    .append("<price>").append("200").append("</price>")
-                    .append("<volume>").append("2000").append("</volume>")
-                    .append("</event>")
-                    .append("</events>");
-            BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(file, true));
-            bufferedWriter.write(sb.toString());
-            bufferedWriter.newLine();
-            bufferedWriter.flush();
-            bufferedWriter.close();
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
+        SiddhiTestHelper.waitForEvents(waitTime, 40, count, timeout);
 
-        siddhiAppRuntime2.restore(snapshot);
-        siddhiAppRuntime2.start();
-        SiddhiTestHelper.waitForEvents(waitTime, 7, count, timeout);
+        File file = new File(dirUri + "/regex/xml");
+        AssertJUnit.assertEquals(0, file.list().length);
+
         //assert event count
-        AssertJUnit.assertEquals("Number of events", 7, count.get());
+        AssertJUnit.assertEquals("Number of events", 40, count.get());
         siddhiAppRuntime.shutdown();
     }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void siddhiIoFileTest17() throws InterruptedException {
+        log.info("test SiddhiIoFile [mode = regex] 17");
+        String streams = "" +
+                "@App:name('TestSiddhiApp')" +
+                "@source(type='file', mode='regex'," +
+                "dir.uri='file:/" + dirUri + "/regex/xml', " +
+                "begin.regex='(*<events>)', " +
+                "tailing='false', " +
+                "@map(type='xml'))" +
+                "define stream FooStream (symbol string, price float, volume long); " +
+                "define stream BarStream (symbol string, price float, volume long); ";
+
+        String query = "" +
+                "from FooStream " +
+                "select * " +
+                "insert into BarStream; ";
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        siddhiAppRuntime.start();
+        Thread.sleep(100);
+        siddhiAppRuntime.shutdown();
+    }
+
 }
