+++ component/src/test/java/io/siddhi/extension/io/file/util/JsonSourceMapper.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.io.file.util;
+package io.siddhi.extension.io.file.util;
 
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonParser;
