+++ pom.xml
@@ -29,7 +29,7 @@
 
     <packaging>pom</packaging>
 
-    <groupId>org.wso2.extension.siddhi.io.file</groupId>
+    <groupId>io.siddhi.extension.io.file</groupId>
     <artifactId>siddhi-io-file-parent</artifactId>
     <version>2.0.1-SNAPSHOT</version>
     <name>Siddhi io file extension</name>
@@ -82,22 +82,22 @@
                 <version>${siddhi.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.xml</groupId>
+                <groupId>io.siddhi.extension.map.xml</groupId>
                 <artifactId>siddhi-map-xml</artifactId>
                 <version>${siddhi.map.xml.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.json</groupId>
+                <groupId>io.siddhi.extension.map.json</groupId>
                 <artifactId>siddhi-map-json</artifactId>
                 <version>${siddhi.map.json.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.text</groupId>
+                <groupId>io.siddhi.extension.map.text</groupId>
                 <artifactId>siddhi-map-text</artifactId>
                 <version>${siddhi.map.text.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.map.csv</groupId>
+                <groupId>io.siddhi.extension.map.csv</groupId>
                 <artifactId>siddhi-map-csv</artifactId>
                 <version>${siddhi.map.csv.version}</version>
             </dependency>
@@ -130,10 +130,10 @@
         <siddhi.version.range>[5.0.0,6.0.0)</siddhi.version.range>
         <wso2.transport.file.version>6.0.52</wso2.transport.file.version>
         <testng.version>6.8</testng.version>
-        <siddhi.map.json.version>5.0.1</siddhi.map.json.version>
-        <siddhi.map.xml.version>5.0.0</siddhi.map.xml.version>
-        <siddhi.map.text.version>2.0.0</siddhi.map.text.version>
-        <siddhi.map.csv.version>2.0.0</siddhi.map.csv.version>
+        <siddhi.map.json.version>5.0.2</siddhi.map.json.version>
+        <siddhi.map.xml.version>5.0.2</siddhi.map.xml.version>
+        <siddhi.map.text.version>2.0.1</siddhi.map.text.version>
+        <siddhi.map.csv.version>2.0.1</siddhi.map.csv.version>
         <jacoco.version>0.7.9</jacoco.version>
         <commons-net.version>3.6</commons-net.version>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
